refactor: feature:map 모듈 내 기능별 패키지 분리
MapScreen을 main/ 패키지로, ScheduleScreen을 schedule/ScheduleBottomSheet로
분리하여 도메인 응집도를 높이고 바텀시트 로직을 별도 Composable로 위임
