+++ feature/map/src/commonMain/kotlin/com/linkit/company/feature/map/main/navigation/MapEntry.kt
@@ -1,9 +1,9 @@
-package com.linkit.company.feature.map.navigation
+package com.linkit.company.feature.map.main.navigation
 
 import androidx.navigation3.runtime.EntryProviderScope
 import androidx.navigation3.runtime.NavKey
 import com.linkit.company.core.navigation.LinkItNavKey
-import com.linkit.company.feature.map.MapScreen
+import com.linkit.company.feature.map.main.MapScreen
 
 fun EntryProviderScope<NavKey>.mapEntry(
     onOpenSchedule: () -> Unit,
